Added a test for missing sub_var.
